+++ Maths/Pollard's Rho.cpp
@@ -3,9 +3,9 @@
 */
 
 namespace PollardRho{
-    long pollardRho(long n) {
-        auto f = [n](long x) { return mul(x, x, n) + 1; };
-        long x = 0, y = 0, t = 30, prd = 2, i = 1, q;
+    long long pollardRho(long long n) {
+        auto f = [n](long long x) { return mul(x, x, n) + 1; };
+        long long x = 0, y = 0, t = 30, prd = 2, i = 1, q;
         while (t++ % 40 || gcd(prd, n) == 1) {
             if (x == y) x = ++i, y = f(x);
             if ((q = mul(prd, max(x,y) - min(x,y), n))) prd = q;
@@ -13,10 +13,10 @@ namespace PollardRho{
         }
         return gcd(prd, n);
     }
-    vector<long> primeFactors(long n) {
+    vector<long long> primeFactors(long long n) {
         if(n == 1) return {};
         if(isPrime(n)) return {n};
-        long x = pollardRho(n);
+        long long x = pollardRho(n);
         auto l = primeFactors(x), r = primeFactors(n / x);
         l.insert(l.end(), all(r));
         return l;
